basic auth added to routes

+++ src/routers/supplies.py
@@ -1,7 +1,6 @@
-
 from fastapi import APIRouter, status, Depends, HTTPException
 from typing import List
-from .. import schemas, models
+from .. import schemas, models, oauth2
 from sqlalchemy.orm import Session
 from .. database import get_db
 
@@ -9,7 +8,7 @@ router = APIRouter(tags=["Supplies"],prefix="/api/supplies")
 
 # GET ALL SUPPLIES
 @router.get("/", status_code=status.HTTP_200_OK, response_model=List[schemas.Supply], tags=["Supplies"])
-async def get_supplies(db: Session = Depends(get_db)):
+async def get_supplies(db: Session = Depends(get_db), current_user:int = Depends(oauth2.get_current_user)):
     """Returns all supplies in the database."""
 
     all_supplies = db.query(models.Supply).all()
@@ -20,7 +19,7 @@ async def get_supplies(db: Session = Depends(get_db)):
 
 # GET A SUPPLY ITEM BY ID
 @router.get("/{id}",status_code=status.HTTP_200_OK, response_model=schemas.Supply, tags=["Supplies"] )
-async def get_single_supply(id:int,db: Session = Depends(get_db)):
+async def get_single_supply(id:int,db: Session = Depends(get_db), current_user:int = Depends(oauth2.get_current_user)):
     """Gets a single supply item from the database based on id."""
 
     supply = db.query(models.Supply).filter(models.Supply.id == id).first()
@@ -35,7 +34,7 @@ async def get_single_supply(id:int,db: Session = Depends(get_db)):
 
 # CREATE A NEW SUPPLY ITEM
 @router.post("/",status_code=status.HTTP_201_CREATED, response_model=schemas.Supply, tags=["Supplies"])
-async def create_new_supply(supply:schemas.CreateSupply, db: Session = Depends(get_db)):
+async def create_new_supply(supply:schemas.CreateSupply, db: Session = Depends(get_db), current_user:int = Depends(oauth2.get_current_user)):
     """Creates a new supply item in the database."""
 
     new_supply_item = models.Supply(**supply.dict())
@@ -48,13 +47,16 @@ async def create_new_supply(supply:schemas.CreateSupply, db: Session = Depends(g
 
 # DELETE A SUPPLY ITEM
 @router.delete("/{id}", tags=["Supplies"])
-async def delete_supply_item(id:int,db: Session = Depends(get_db)):
+async def delete_supply_item(id:int,db: Session = Depends(get_db), current_user:int = Depends(oauth2.get_current_user)):
     """Deletes a supply item from the database based on id."""
 
     supply_item = db.query(models.Supply).filter(models.Supply.id == id).first()
 
     if supply_item == None:
             raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"User with id: {id} was not found")
+
+    if current_user != supply_item.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to perform requested action.")
 
     db.delete(supply_item)
     db.commit()
@@ -63,7 +65,7 @@ async def delete_supply_item(id:int,db: Session = Depends(get_db)):
 
 # UPDATE A SUPPLY ITEM - ALL ATTRIBUTES
 @router.put("/{id}", tags=["Supplies"], response_model=schemas.Supply)
-async def update_supply_item(id:int, new_supply_data:schemas.UpdateSupply,db: Session = Depends(get_db)):
+async def update_supply_item(id:int, new_supply_data:schemas.UpdateSupply,db: Session = Depends(get_db), current_user:int = Depends(oauth2.get_current_user)):
 
     query_item = db.query(models.Supply).filter(models.Supply.id == id)
